+++ src/components/history-list.tsx
@@ -1,7 +1,7 @@
 import { format, parseISO } from 'date-fns';
 import React, { Fragment } from 'react';
 // No longer import Event directly, T will be generic
-import { useSortedEvents } from '@/hooks/useSortedEvents';
+import { useSortedEvents } from '@/hooks/use-sorted-events';
 
 interface ItemWithId {
 	id: string;
